requirements.txt
# Core Flask and Web Framework
Flask==2.3.3
gunicorn==21.2.0
Werkzeug==2.3.7

# Video Processing
moviepy==1.0.3
imageio==2.31.5
imageio-ffmpeg==0.4.9

# Image Processing
Pillow==10.0.1
opencv-python==4.8.1.78

# Audio Processing and TTS
coqui-tts==0.17.8
pyttsx3==2.90
torch==2.1.0
torchaudio==2.1.0
soundfile==0.12.1

# Web Scraping and Content Processing
newspaper3k==0.2.8
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
readability==0.3.1

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3
scikit-learn==1.3.2
wordcloud==1.9.2

# Data Processing
pandas==2.1.3
numpy==1.24.4
PyYAML==6.0.1

# Development and Utilities
python-dotenv==1.0.0
pathlib2==2.3.7
python-dateutil==2.8.2

# Optional: For enhanced functionality
# Uncomment if needed
# openai==0.28.1
# transformers==4.35.2
# diffusers==0.21.4